API: unify update to generic.py
